+++ gst/gsturi.c
@@ -268,7 +268,7 @@ unescape_string (const gchar * escaped_string, const gchar * illegal_characters)
   }
 
   *out = '\0';
-  g_assert (out - result <= strlen (escaped_string));
+  g_assert ((size_t) (out - result) <= strlen (escaped_string));
   return result;
 
 }
